+++ WebDriverAgentLib/Utilities/FBScreenshot.m
@@ -123,7 +123,7 @@ NSString *formatTimeInterval(NSTimeInterval interval) {
   return screenshotData;
 }
 
-+ (NSData *)takeWithScreenID:(unsigned int)screenID
++ (NSData *)takeWithScreenID:(long long)screenID
                        scale:(CGFloat)scale
           compressionQuality:(CGFloat)compressionQuality
                         rect:(CGRect)rect
@@ -146,7 +146,7 @@ NSString *formatTimeInterval(NSTimeInterval interval) {
                                                         error:error];
 }
 
-+ (NSData *)takeInOriginalResolutionWithScreenID:(unsigned int)screenID
++ (NSData *)takeInOriginalResolutionWithScreenID:(long long)screenID
                               compressionQuality:(CGFloat)compressionQuality
                                              uti:(NSString *)uti
                                          timeout:(NSTimeInterval)timeout
@@ -254,7 +254,7 @@ NSString *formatTimeInterval(NSTimeInterval interval) {
   return imageEncoding;
 }
 
-+ (nullable id)screenshotRequestWithScreenID:(unsigned int)screenID
++ (nullable id)screenshotRequestWithScreenID:(long long)screenID
                                         rect:(struct CGRect)rect
                                          uti:(NSString *)uti
                           compressionQuality:(CGFloat)compressionQuality
